# odahuFlow/operator/docs/swagger.yaml
definitions:
...
  HTTPResult:
    properties:
      message:
        description: Success of error message
        type: string
    type: object
...
  ModelTraining:
    properties:
      id:
        description: Model training ID
        type: string
      spec:
        $ref: '#/definitions/ModelTrainingSpec'
        description: Model training specification
        type: object
      status:
        $ref: '#/definitions/ModelTrainingStatus'
        description: Model training status
        type: object
    type: object
...
  DataBindingDir:
    properties:
      connName:
        description: Connection name for data
        type: string
      localPath:
        description: Local training path
        type: string
      remotePath:
        description: Overwrite remote data path in connection
        type: string
    type: object
  EnvironmentVariable:
    properties:
      name:
        description: Name of an environment variable
        type: string
      value:
        description: Value of an environment variable
        type: string
    type: object
...
  ModelIdentity:
    properties:
      artifactNameTemplate:
        description: Template of output artifact name
        type: string
      name:
        description: Model name
        type: string
      version:
        description: Model version
        type: string
    type: object
...
  ModelTrainingSpec:
    properties:
      args:
        items:
          type: string
        type: array
      data:
        description: Input data for a training
        items:
          $ref: '#/definitions/DataBindingDir'
        type: array
      entrypoint:
        description: Model training file. It can be python\bash script or jupiter
          notebook
        type: string
      envs:
        description: Custom environment variables that should be set before entrypoint
          invocation.
        items:
          $ref: '#/definitions/EnvironmentVariable'
        type: array
      hyperParameters:
        description: Model training hyperParameters in parameter:value format
        type: object
      image:
        description: Train image
        type: string
      model:
        $ref: '#/definitions/ModelIdentity'
        description: Model Identity
        type: object
      reference:
        description: VCS Reference
        type: string
      resources:
        $ref: '#/definitions/ResourceRequirements'
        description: |-
          Resources for model container
          The same format like k8s uses for pod resources.
        type: object
      toolchain:
        description: IntegrationName of toolchain
        type: string
      vcsName:
        description: Name of Connection resource. Must exists
        type: string
      workDir:
        description: Directory with model scripts/files in a git repository
        type: string
    type: object
  ModelTrainingStatus:
    properties:
      artifacts:
        description: List of training results
        items:
          $ref: '#/definitions/TrainingResult'
        type: array
      exitCode:
        description: Pod exit code
        type: integer
      message:
        description: Pod last log
        type: string
      podName:
        description: Pod package for name
        type: string
      reason:
        description: Pod reason
        type: string
      state:
        description: Model Packaging State
        type: string
    type: object
...
  ResourceList:
    properties:
      cpu:
        description: Read more about CPU resource here https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/#meaning-of-cpu
        type: string
      gpu:
        description: Read more about GPU resource here https://kubernetes.io/docs/tasks/manage-gpus/scheduling-gpus/#using-device-plugins
        type: string
      memory:
        description: Read more about memory resource here https://kubernetes.io/docs/concepts/configuration/manage-compute-resources-container/#meaning-of-memory
        type: string
    type: object
  ResourceRequirements:
    properties:
      limits:
        $ref: '#/definitions/ResourceList'
        description: Limits describes the maximum amount of compute resources allowed.
        type: object
      requests:
        $ref: '#/definitions/ResourceList'
        description: Requests describes the minimum amount of compute resources required.
        type: object
    type: object
...
  TrainingResult:
    properties:
      artifactName:
        description: Trained artifact name
        type: string
      commitID:
        description: VCS commit
        type: string
      runId:
        description: Mlflow run ID
        type: string
    type: object
info:
  contact: {}
  description: This is an API Gateway server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Gateway
  version: "1.0"
paths:
...
  /api/v1/model/training:
    get:
      consumes:
      - application/json
      description: Get list of Model Trainings
      parameters:
      - description: Number of entities in a response
        in: path
        name: size
        type: integer
      - description: Number of a page
        in: path
        name: page
        type: integer
      - description: Model name
        in: path
        name: model_name
        type: integer
      - description: Model version
        in: path
        name: model_version
        type: integer
      - description: Toolchain name
        in: path
        name: toolchain
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ModelTraining'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get list of Model Trainings
      tags:
      - Training
    post:
      consumes:
      - application/json
      description: Create a Model Training. Results is created Model Training.
      parameters:
      - description: Create a Model Training
        in: body
        name: mt
        required: true
        schema:
          $ref: '#/definitions/ModelTraining'
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ModelTraining'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Create a Model Training
      tags:
      - Training
    put:
      consumes:
      - application/json
      description: Update a Model Training. Results is updated Model Training.
      parameters:
      - description: Update a Model Training
        in: body
        name: mt
        required: true
        schema:
          $ref: '#/definitions/ModelTraining'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ModelTraining'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Update a Model Training
      tags:
      - Training
  /api/v1/model/training/{id}:
    delete:
      consumes:
      - application/json
      description: Get a Model Training by id
      parameters:
      - description: Model Training id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HTTPResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get a Model Training
      tags:
      - Training
    get:
      consumes:
      - application/json
      description: Get a Model Training by id
      parameters:
      - description: Model Training id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ModelTraining'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Get a Model Training
      tags:
      - Training
  /api/v1/model/training/{id}/log:
    get:
      consumes:
      - text/plain
      description: Stream logs from model training pod
      parameters:
      - description: follow logs
        in: query
        name: follow
        type: boolean
      - description: Model Training id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Stream logs from model training pod
      tags:
      - Training
  /api/v1/model/training/{id}/result:
    put:
      consumes:
      - application/json
      description: Save a Model Training by id
      parameters:
      - description: Model Training result
        in: body
        name: MP
        required: true
        schema:
          $ref: '#/definitions/TrainingResult'
          type: object
      - description: Model Training id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TrainingResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTPResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTPResult'
      summary: Save a Model Training result
      tags:
      - Training
...
swagger: "2.0"